Basic refactoring of views
