+++ app/data/1_0205.json
@@ -1,10 +1,14 @@
 {
-    "minme_type": "application/json/ons/eq",
+    "mime_type": "application/json/ons/eq",
     "schema_version": "0.0.1",
     "questionnaire_id": "23",
     "survey_id": "23",
     "title": "Monthly Business Survey - Retail Sales Index",
     "description": "MCI description",
+    "introduction": {
+      "legal":"<b>You are required by law to complete this questionnaire.</b> If you do not complete and submit this questionnaire by 6th February 2016, penalties may be incurred (under section 4 of the Statistics of Trade Act 1947). All the information you provide is kept strictly confidential. It is illegal for us to reveal your data or identify your business to unauthorised persons.",
+      "description": "<p>The information supplied is used to produce monthly estimates of the total retail sales in Great Britain where retailing is defined as the sale of goods to the general public for household consumption. The Retail Sales Index is a key indicator of the progress of the economy. It is also used to help estimate consumer spending on retail goods and the output of the retail sector, both of which feed into the compilation of the National Accounts. The results are also used by the Bank of England and HM Treasury to inform decision making by government and in formulating financial policies.</p><p>We guarantee that while your employment is less than 10, you will receive no more than 15 monthly questionnaires for this one ONS business survey. You must complete and return all questionnaires on time, after which you will be excluded from all business surveys for at least 3 years. The Annual Survey of Hours and Earnings is not covered by this guarantee.</p>"
+    },
     "groups": [
         {
             "id": "14ba4707-321d-441d-8d21-b8367366e766",
