+++ app/main/__init__.py
@@ -3,6 +3,9 @@ from flask import Blueprint
 main_blueprint = Blueprint('main', __name__)
 
 from app.main.views import root         # NOQA
+from app.main.views import questionnaire  # NOQA
+from app.main.views import cover_page   # NOQA
+from app.main.views import thank_you    # NOQA
 
 
 @main_blueprint.after_request
