+++ app/templates/cover-page.html
@@ -47,6 +47,70 @@
       </div>
     </section>
 
+
+
+    <div class="wrapper">
+
+        <div class="masthead">
+          <div class="logo">
+            <img src="img/logo.svg">
+          </div>
+          <div class="survey-info">
+            <div class="survey-icon"><img src="img/info.svg"></div>
+            <ul class="info-group">
+              <li class="info-block">Survey Code: 023</li>
+              <li class="info-block">Period: 201601</li>
+              <li class="info-block-last">Reference: 4990 0000 000</li>
+            </ul>
+          </div>
+          <div class="support">
+            <a href="/"><img src="img/help.svg">Need help?</a>
+          </div>
+        </div>
+
+        <main>
+          <div class="survey-user-info">
+            <div class="address-data">
+              <b>For the attention of</b>
+              <address>
+                CONTACT NAME<br>
+                OFFICE FOR NATIONAL STATISTICS<br>
+                GOVERNMENT BUILDINGS<br>
+                CARDIFF ROAD<br>
+                NEWPORT<br>
+                NP10 8XG
+              </address>
+              <p><a href=""><small>Details correct?</small></a></p>
+            </div>
+            <div class="survey-notice">
+              <h5>Please submit by</h5>
+              <p><date>6th February 2016</date></p>
+              <h5>Period</h5>
+              <p class="period">4/1/2016 to 31/1/2016 (4 Weeks)</p>
+              <p class="attn">NB: Your response is legally required</p>
+            </div>
+          </div>
+          <h2>Monthly Business Survey - Retail Sales Index</h2>
+          <h5>Notice is given under section 1 of the Statistics of Trade Act 1947</h5>
+          <div class="intro-text">
+            <p><b>You are required by law to complete this questionnaire.</b> If you do not complete and submit this questionnaire by 6th February 2016, penalties may be incurred (under section 4 of the Statistics of Trade Act 1947). All the information you provide is kept strictly confidential. It is illegal for us to reveal your data or identify your business to unauthorised persons.</p>
+
+            <p class="callout"><i>i</i> If actual figures are not available, please provide informed estimates.</p>
+
+            <div class="panels">
+                <input type="checkbox" id="instructions">
+                <label for="instructions"><span></span>How is your information used?</label>
+                <article>
+                  <p>The information supplied is used to produce monthly estimates of the total retail sales in Great Britain where retailing is defined as the sale of goods to the general public for household consumption. The Retail Sales Index is a key indicator of the progress of the economy. It is also used to help estimate consumer spending on retail goods and the output of the retail sector, both of which feed into the compilation of the National Accounts. The results are also used by the Bank of England and HM Treasury to inform decision making by government and in formulating financial policies.</p>
+                  <p>We guarantee that while your employment is less than 10, you will receive no more than 15 monthly questionnaires for this one ONS business survey. You must complete and return all questionnaires on time, after which you will be excluded from all business surveys for at least 3 years. The Annual Survey of Hours and Earnings is not covered by this guarantee.</p>
+                </article>
+            </div>
+
+
+            <p><b>Office for National Statistics</b></p>
+
+
+
     <footer class="cover__footer">
 
       <a href="/questionnaire" class="btn">Get Started</a>
